updated build

+++ build.cmd
@@ -3,6 +3,22 @@ setlocal enabledelayedexpansion
 
 set PATH="C:\Program Files\7-Zip";%PATH%
 
+where /q git.exe || (
+  echo ERROR: "git.exe" not found
+  exit /b 1
+)
+
+if exist "%ProgramFiles%\7-Zip\7z.exe" (
+  set SZIP="%ProgramFiles%\7-Zip\7z.exe"
+) else (
+  where /q 7za.exe || (
+    echo ERROR: 7-Zip installation or "7za.exe" not found
+    exit /b 1
+  )
+  set SZIP=7za.exe
+)
+
+
 rem *** Visual Studio environment ***
 
 where /Q cl.exe || (
@@ -20,43 +36,49 @@ rem *** ninja
 
 where /q ninja.exe || (
   curl -LOsf https://github.com/ninja-build/ninja/releases/download/v1.11.0/ninja-win.zip || exit /b 1
-  7z x -bb0 -y ninja-win.zip 1>nul 2>nul || exit /b 1
+  %SZIP% x -bb0 -y ninja-win.zip 1>nul 2>nul || exit /b 1
   del ninja-win.zip 1>nul 2>nul
 )
 
 
-rem *** fetching latest release version
+rem *** fetch latest release version
 
 for /F "delims=" %%v in ('"curl -sfL https://api.github.com/repos/aseprite/aseprite/releases/latest | jq .tag_name -r"') do (
   set ASEPRITE_VERSION=%%v
 )
+echo building %ASEPRITE_VERSION%
 
 
-rem **** checking if release is already built
+rem **** clone aseprite repo
 
-curl -sfLo nul https://api.github.com/repos/mmozeiko/aseprite-bin/releases/tags/%ASEPRITE_VERSION%
-if %ERRORLEVEL% EQU 0 (
-  echo release already exists, exiting
-  exit /b 0
+if exist aseprite (
+  pushd aseprite
+  call git clean --quiet -fdx
+  call git submodule foreach --recursive git clean -xfd
+  call git fetch --quiet --depth=1 --no-tags origin %ASEPRITE_VERSION%:refs/remotes/origin/%ASEPRITE_VERSION% || echo "failed to fetch latest version"     && exit /b 1
+  call git reset --quiet --hard origin/%ASEPRITE_VERSION%                                                     || echo "failed to update to latest version" && exit /b 1
+  call git submodule update --init --recursive                                                                || echo "failed to update submodules"        && exit /b 1
+  popd
+) else (
+  call git clone --quiet -c advice.detachedHead=false --no-tags --recursive --depth=1 -b "%ASEPRITE_VERSION%" https://github.com/aseprite/aseprite.git || echo "failed to clone repo" && exit /b 1
 )
-
-
-rem **** cloning asesprite repo
-
-git clone --quiet -c advice.detachedHead=false --no-tags --recursive --depth=1 -b "%ASEPRITE_VERSION%" https://github.com/aseprite/aseprite.git || echo "failed to clone repo" && exit /b 1
 python -c "v = open('aseprite/src/ver/CMakeLists.txt').read(); open('aseprite/src/ver/CMakeLists.txt', 'w').write(v.replace('1.x-dev', '%ASEPRITE_VERSION%'[1:]))"
 
 
-rem *** downloading skia
+rem *** download skia
 
-mkdir skia
-cd skia
-curl -sfLO https://github.com/aseprite/skia/releases/download/m102-861e4743af/Skia-Windows-Release-x64.zip || echo failed to download skia && exit /b 1
-7z x -y Skia-Windows-Release-x64.zip
-cd ..
+if not exist skia (
+  mkdir skia
+  pushd skia
+  curl -sfLO https://github.com/aseprite/skia/releases/download/m102-861e4743af/Skia-Windows-Release-x64.zip || echo failed to download skia && exit /b 1
+  7z x -y Skia-Windows-Release-x64.zip
+  popd
+)
 
 
-rem *** building asesprite
+rem *** build aseprite
+
+if exist build rd /s /q build
 
 set LINK=opengl32.lib
 cmake                                          ^
@@ -64,8 +86,10 @@ cmake                                          ^
   -S aseprite                                  ^
   -B build                                     ^
   -DCMAKE_BUILD_TYPE=Release                   ^
-  -DCMAKE_C_FLAGS="/MP"                        ^
-  -DCMAKE_CXX_FLAGS="/MP"                      ^
+  -DCMAKE_POLICY_DEFAULT_CMP0074=NEW           ^
+  -DCMAKE_POLICY_DEFAULT_CMP0091=NEW           ^
+  -DCMAKE_POLICY_DEFAULT_CMP0092=NEW           ^
+  -DCMAKE_MSVC_RUNTIME_LIBRARY=MultiThreaded   ^
   -DENABLE_CCACHE=OFF                          ^
   -DOPENSSL_USE_STATIC_LIBS=TRUE               ^
   -DLAF_BACKEND=skia                           ^
@@ -75,14 +99,16 @@ cmake                                          ^
 ninja -C build || echo "build failed" && exit /b 1
 
 
-rem *** creating zip file
+rem *** create output folder
 
 mkdir aseprite-%ASEPRITE_VERSION%
 echo # This file is here so Aseprite behaves as a portable program >aseprite-%ASEPRITE_VERSION%\aseprite.ini
 xcopy /E /Q /Y aseprite\docs aseprite-%ASEPRITE_VERSION%\docs\
 xcopy /E /Q /Y build\bin\aseprite.exe aseprite-%ASEPRITE_VERSION%\
 xcopy /E /Q /Y build\bin\data aseprite-%ASEPRITE_VERSION%\data\
-7z a -r aseprite-%ASEPRITE_VERSION%.zip aseprite-%ASEPRITE_VERSION% || echo failed to create output zip file && exit /b
 
-
-echo ::set-output name=ASEPRITE_VERSION::%ASEPRITE_VERSION%
+if "%GITHUB_WORKFLOW%" neq "" (
+  mkdir github
+  move aseprite-%ASEPRITE_VERSION% github\
+  echo ASEPRITE_VERSION=%ASEPRITE_VERSION%>>%GITHUB_OUTPUT%
+)
